requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
dataclasses-json==0.5.14
diskcache==5.6.3
frozenlist==1.4.0
greenlet==2.0.2
idna==3.4
langchain==0.0.292
langsmith==0.0.37
llama-cpp-python==0.2.6
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.6
numpy==1.24.4
packaging==23.1
pika==1.3.2
pydantic==2.3.0
pydantic-core==2.6.3
PyMuPDF==1.23.3
PyMuPDFb==1.23.3
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
SQLAlchemy==2.0.20
tenacity==8.2.3
tiktoken==0.5.1
typing-extensions==4.7.1
typing-inspect==0.9.0
urllib3==2.0.4
yarl==1.9.2
